fix: revert breaking change for `compiler/preprocess` types location
Reverts the change in `interface` locations in abf11bb02b2afbd3e4cac509a0f70e318c306364
No functionality changes.
